requirements.txt
# requirements.txt
pyinstaller
selenium
python-dotenv
psutil
pyautogui
pillow
opencv-python
pygetwindow